pyproject.toml
[project]
name = "soccer-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "opencv-python-headless>=4.13.0.92",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "reflex>=0.9.0",
    "sqlalchemy[asyncio]>=2.0.49",
    "sse-starlette>=3.3.4",
    "uvicorn[standard]>=0.46.0",
]
